Conflicts:
	src/apis/c/nanos.h
	src/apis/c/nanos_wd.cpp
	src/core/basethread_decl.hpp
	src/core/nanos-int.h
	src/core/schedule.cpp
	src/core/system.cpp
	src/core/system_decl.hpp
	src/plugins/sched/affinity_sched.cpp
	src/support/memtracker.hpp
	tests/test/api/c/sort.c
	tests/test/core/deps/test_dependencies.c
	tests/test/core/deps/test_dependency_types.c

+++ src/core/workdescriptor.hpp
@@ -49,8 +49,8 @@ inline bool WorkDescriptor::started ( void ) const { return (( _state != INIT ) 
 inline size_t WorkDescriptor::getDataSize () const { return _data_size; }
 inline void WorkDescriptor::setDataSize ( size_t data_size ) { _data_size = data_size; }
 
-inline int WorkDescriptor::getDataAlignment () const { return _data_align; }
-inline void WorkDescriptor::setDataAlignment ( int data_align ) { _data_align = data_align; }
+inline size_t WorkDescriptor::getDataAlignment () const { return _data_align; }
+inline void WorkDescriptor::setDataAlignment ( size_t data_align ) { _data_align = data_align; }
 
 inline WorkDescriptor * WorkDescriptor::getParent() { return _parent; }
 inline void WorkDescriptor::setParent ( WorkDescriptor * p ) { _parent = p; }
